requirements.txt
geopandas==0.6.1
loguru==0.4.0
numpy==1.18.1
pytest==5.4.1
requests==2.22.0
rtree==0.8.3
pyproj==2.4.1
pyyaml==5.4.1

